LayoutTests/imported/w3c/web-platform-tests/fetch/corb/style-html-correctly-labeled.sub-expected.txt

CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/fetch/corb/resources/html-correctly-labeled.html
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/fetch/corb/resources/html-correctly-labeled.html
Header example

Paragraph body


PASS Untitled
